% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANOVA_of_summary2.R
\name{ANOVA_of_summary2}
\alias{ANOVA_of_summary2}
\title{Compare 3 or more groups using ANOVA and only the means + stds}
\usage{
ANOVA_of_summary2(groups, means, stds, var.equal = T)
}
\description{
Compare 3 or more groups using ANOVA and only the means + stds
}
\examples{
ANOVA_of_summary2()
}
\keyword{data_processing}
