% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box_mets.R
\name{box_mets}
\alias{box_mets}
\title{Create beautiful boxplots from data.}
\usage{
box_mets(
  dataset,
  id_mets,
  x_variable,
  fill_variable,
  common_legend = T,
  stat_test = F,
  ref_group,
  y_title
)
}
\description{
Create beautiful boxplots from data.
}
\examples{
box_mets()
}
\keyword{data_processing}
