% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxMets.R
\name{boxMets}
\alias{boxMets}
\title{Create beautiful boxplots from data.}
\usage{
boxMets(
  dataset,
  id_mets,
  x_variable,
  fill_variable,
  common_legend = T,
  stat_test = F,
  ref_group,
  y_title
)
}
\description{
Create beautiful boxplots from data.
}
\examples{
boxMets()
}
\keyword{data_processing}
