% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_cell_growth.R
\name{calculate_cell_growth}
\alias{calculate_cell_growth}
\title{Get growth rate from cell numbers and time}
\usage{
calculate_cell_growth(cell_no, timepoints)
}
\description{
Get growth rate from cell numbers and time
}
\examples{
calculate_cell_growth()
}
\keyword{Cells}
