% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateCellGrowth.R
\name{calculateCellGrowth}
\alias{calculateCellGrowth}
\title{Get growth rate from cell numbers and time}
\usage{
calculateCellGrowth(cell_no, timepoints)
}
\description{
Get growth rate from cell numbers and time
}
\examples{
calculateCellGrowth()
}
\keyword{Cells}
