% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractCorrelationMatrix.R
\name{extractCorrelationMatrix}
\alias{extractCorrelationMatrix}
\title{A function to identify genes in a gene expression matrix. The dataset needs to have genes in columns and patients/cells in rows}
\usage{
extractCorrelationMatrix(dataset)
}
\description{
A function to identify genes in a gene expression matrix. The dataset needs to have genes in columns and patients/cells in rows
}
\examples{
extractCorrelationMatrix()
}
\keyword{data_processing}
