% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractDEgenesTCGA.R
\name{extractDEgenesTCGA}
\alias{extractDEgenesTCGA}
\title{Differential gene analysis}
\usage{
extractDEgenesTCGA(group1, group2, fc, pval, return_val)
}
\description{
A function for extracting DE genes from TCGA data. This is assuming the data is normalized and uses
students t-test to evaluate, and BH-FDR correction. FC refers to log-fold change...
group1 and group2 refer to different cluster of samples from same dataframe. 'return_val' tells you
whether to take the significantly higher genes from group1 or group2 (1 and 2, respectively).
}
\examples{
extractDEgenesTCGA()
}
\keyword{DEgenes}
