% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findBestSurvival.R
\name{findBestSurvival}
\alias{findBestSurvival}
\title{Extracting the best split of survival of patients from expression of genes.}
\usage{
findBestSurvival(dataset, genelist, quantile_list, cor_data,
  print_plot = TRUE, effect_type = TRUE)
}
\description{
A function for optimizing the value chosen for survival signature generation
}
\examples{
findBestSurvival()
}
\keyword{survival}
