% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hierarchical_clustering.R
\name{Hierarchical_clustering}
\alias{Hierarchical_clustering}
\title{Hierarchical clustering of dataframe}
\usage{
Hierarchical_clustering(dataframe, gene_id, cor_method)
}
\description{
A function for Hierarchical clustering (Spearman or Pearson) of data from cBioPortal (raw data).
Useful for looking at specific patients or all of them.
}
\examples{
Hierarchical_clustering()
}
\keyword{clustering}
