% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_installed.R
\name{is_installed}
\alias{is_installed}
\title{Installation check of packages}
\usage{
is_installed(mypkg)
}
\description{
This function allows you to check whether a certain package is installed or not
}
\examples{
is_installed('ggplot2')
}
\keyword{packages}
