% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_or_install.R
\name{load_or_install}
\alias{load_or_install}
\title{Installation check of packages}
\usage{
load_or_install(package_names)
}
\description{
A function for providing the necessary packages for the functions it uses...
}
\examples{
load_or_install('ggplot2')
}
\keyword{packages}
