% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metPCA.R
\name{metPCA}
\alias{metPCA}
\title{PCA plotting of data:}
\usage{
metPCA(
  dataset,
  groups,
  scale,
  pcs,
  center,
  log,
  power,
  colors,
  CI,
  label,
  corplot,
  minvalue,
  narm,
  scree,
  nh,
  pointsize,
  corrList,
  loadinglist
)
}
\description{
metPCA from Freyr Johannson.
}
\examples{
metPCA()
}
\keyword{PCA}
