% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_max.R
\name{min_max}
\alias{min_max}
\title{Function to normalize data (good in 'apply' of rows)}
\usage{
min_max(x, a, b)
}
\description{
Function to normalize data (good in 'apply' of rows)
}
\examples{
min_max(x,a,b)
}
\keyword{Normalization}
