% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/not_in.R
\name{not_in}
\alias{not_in}
\title{Function to extract all parts of dataframe_whole EXCEPT dataframe_part}
\usage{
not_in(dataframe_whole, dataframe_part)
}
\description{
Function to extract all parts of dataframe_whole EXCEPT dataframe_part
}
\keyword{intersect}
