% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperfunctions.R, R/nth_highest_val.R
\name{nth_highest_val}
\alias{nth_highest_val}
\title{Helper functions for stuff...}
\usage{
nth_highest_val(my_vector, n)

nth_highest_val(my_vector, n)
}
\description{
Helper functions for stuff...

Function to extract the n-th highest value in a vector
}
\examples{
nth_highest_val(my_vector,n)
}
\keyword{indexing}
\keyword{various}
