% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_gain_analysis.R
\name{p_gain_analysis}
\alias{p_gain_analysis}
\title{# A quite raw function which can find ratios that are more interesting than single pairs of predictors in lin.reg.models}
\usage{
p_gain_analysis(dataset, covariates, start, p_thresh)
}
\description{
# A quite raw function which can find ratios that are more interesting than single pairs of predictors in lin.reg.models
}
\examples{
p_gain_analysis()
}
\keyword{data_processing}
