% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_look_for_paper.R
\name{plot_look_for_paper}
\alias{plot_look_for_paper}
\title{Making plots look good}
\usage{
plot_look_for_paper(
  p,
  y_text,
  x_text,
  rotate_check = FALSE,
  legend_check = TRUE
)
}
\description{
Making plots look good
}
\examples{
plot_look_for_paper()
}
\keyword{plots}
