% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_metabolites.R
\name{plot_metabolites}
\alias{plot_metabolites}
\title{A function to plot multiple plots from a list to the same graph, with a common legend or not}
\usage{
plot_metabolites(plots_metabolites, nplots, common_legend = T)
}
\description{
A function to plot multiple plots from a list to the same graph, with a common legend or not
}
\examples{
plot_metabolites()
}
\keyword{data_processing}
