% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLookForPaper.R
\name{plotLookForPaper}
\alias{plotLookForPaper}
\title{Making plots look good}
\usage{
plotLookForPaper(p, y_text, x_text, rotate_check = FALSE,
  legend_check = TRUE)
}
\description{
Making plots look good
}
\examples{
plotLookForPaper()
}
\keyword{plots}
