% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareRNAdata.R
\name{prepareRNAdata}
\alias{prepareRNAdata}
\title{Preparing RNA data for PCA plotting}
\usage{
prepareRNAdata(rna_data, clinical_data, what_to_check, gene_identifier)
}
\description{
Function for preparing the RNA data for metPCA acc. to protocol.
rna_data' is the raw RNAseq data from cBioPortal. Clinical data is the corresponding information on patients.
groups is how many groups the patients should be split into based on survival (can be changed)
}
\examples{
prepareRNAdata()
}
\keyword{data_processing}
