% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareSurvivalDataTCGA.R
\name{prepareSurvivalDataTCGA}
\alias{prepareSurvivalDataTCGA}
\title{Preparing data for survival analysis}
\usage{
prepareSurvivalDataTCGA(dataset, genes, quantile_used, percentage_patients)
}
\description{
A function to prepare data (genes of interest) for survival analysis.
}
\examples{
prepareSurvivalDataTCGA()
}
\keyword{data_processing}
