% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Raw2MDV.R
\name{Raw2MDV}
\alias{Raw2MDV}
\title{Function to normalize a vector so that 1) No values are below tol_value and 2) total sum of vector is 1}
\usage{
Raw2MDV(iso_vector, tol_value)
}
\description{
Function to normalize a vector so that 1) No values are below tol_value and 2) total sum of vector is 1
}
\examples{
Raw2MDV()
}
\keyword{data_processing}
