% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_2_mdv.R
\name{raw_2_mdv}
\alias{raw_2_mdv}
\title{Function to normalize a vector so that 1) No values are below tol_value and 2) total sum of vector is 1}
\usage{
raw_2_mdv(iso_vector, tol_value)
}
\description{
Function to normalize a vector so that 1) No values are below tol_value and 2) total sum of vector is 1
}
\examples{
raw_2_mdv()
}
\keyword{data_processing}
