% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_low_variance.R
\name{remove_low_variance}
\alias{remove_low_variance}
\title{A function to remove rows with zero variance (where sd is zero)}
\usage{
remove_low_variance(dataset, threshold, rows_or_cols)
}
\description{
A function to remove rows with zero variance (where sd is zero)
}
\examples{
remove_low_variance()
}
\keyword{data_processing}
