% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_outliers.R
\name{remove_outliers}
\alias{remove_outliers}
\title{Remove outliers from vector above 1.25 x IQR}
\usage{
remove_outliers(x, na.rm = TRUE, ...)
}
\description{
Remove outliers from vector above 1.25 x IQR
}
\examples{
remove_outliers()
}
\keyword{data_processing}
