% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeLowVariance.R
\name{removeLowVariance}
\alias{removeLowVariance}
\title{A function to remove rows with zero variance (where sd is zero)}
\usage{
removeLowVariance(dataset, threshold, rows_or_cols)
}
\description{
A function to remove rows with zero variance (where sd is zero)
}
\examples{
removeLowVariance()
}
\keyword{data_processing}
