% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROC_maker.R
\name{ROC_maker}
\alias{ROC_maker}
\title{Create beautiful boxplots from data.}
\usage{
ROC_maker(houston_test, return_val)
}
\description{
Create beautiful boxplots from data.
}
\examples{
ROC_maker()
}
\keyword{plots}
