% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_vector.R
\name{sample_vector}
\alias{sample_vector}
\title{A function for sampling n rows from dataframe.}
\usage{
sample_vector(vector, n, x, replicate = T)
}
\description{
A function for sampling n rows from dataframe.
}
\examples{
sample_vector()
}
\keyword{data_processing}
