% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sem.R
\name{sem}
\alias{sem}
\title{Function to output the standard error of mean from input vector}
\usage{
sem(x)
}
\description{
Function to output the standard error of mean from input vector
}
\examples{
sem()
}
\keyword{data_analysis}
