% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.R
\name{standardize}
\alias{standardize}
\title{Function to standardize dataframe (potentially more global than 'standardizeTCGA)}
\usage{
standardize(data_frame)
}
\description{
Function to standardize dataframe (potentially more global than 'standardizeTCGA)
}
\keyword{Normalization}
