% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardizeTCGA.R
\name{standardizeTCGA}
\alias{standardizeTCGA}
\title{Standardizing TCGA data}
\usage{
standardizeTCGA(data)
}
\description{
A function to standardize data (x-mean/std)
}
\examples{
standardizeTCGA()
}
\keyword{data_processing}
