% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_cutpoint_multiple.R
\name{surv_cutpoint_multiple}
\alias{surv_cutpoint_multiple}
\title{# A function to check whether multiple genes can be used to discriminate groups in terms of survival curves}
\usage{
surv_cutpoint_multiple(
  meta_data,
  features,
  time,
  event,
  minprop,
  format_signature
)
}
\description{
# A function to check whether multiple genes can be used to discriminate groups in terms of survival curves
}
\examples{
surv_cutpoint_multiple()
}
\keyword{survival_analysis}
