% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttest2.R
\name{t.test2}
\alias{t.test2}
\title{A function to perform a two-sample t-test based on summary, not replicate values.}
\usage{
\method{t}{test2}(m1, m2, s1, s2, n1, n2, m0 = 0,
  equal.variance = FALSE)
}
\description{
A function to perform a two-sample t-test based on summary, not replicate values.
}
\examples{
t.test2()
}
\keyword{data_processing}
