% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_match_order.R
\name{test_match_order}
\alias{test_match_order}
\title{Function to check whether two vectors are in the same order or not.}
\usage{
test_match_order(x, y)
}
\description{
Function to check whether two vectors are in the same order or not.
}
\examples{
test_match_order()
}
\keyword{data_processing}
