% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEM.R
\name{SEM}
\alias{SEM}
\title{Function to output the standard error of mean from input vector}
\usage{
SEM(x)
}
\description{
Function to output the standard error of mean from input vector
}
\examples{
SEM()
}
\keyword{data_analysis}
